*** empty log message ***

+++ lib/parser-sming.tab.c
@@ -361,7 +361,7 @@ checkDate(Parser *parserPtr, char *date)
     if (len == 10 || len == 16) {
 	for (i = 0; i < len; i++) {
 	    if (((i < 4 || i == 5 || i == 6 || i == 8 || i == 9 || i == 11
-		  || i == 12 || i == 14 || i == 15) && ! isdigit(date[i]))
+		  || i == 12 || i == 14 || i == 15) && ! isdigit((int)date[i]))
 		|| ((i == 4 || i == 7) && date[i] != '-')
 		|| (i == 10 && date[i] != ' ')
 		|| (i == 13 && date[i] != ':')) {
